requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
psycopg2-binary==2.9.10
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0